Update - return type

+++ 0x01-python_async_function/0-basic_async_syntax.py
@@ -4,7 +4,7 @@ import random
 import asyncio
 
 
-async def wait_random(max_delay: int = 0) -> int:
+async def wait_random(max_delay: int = 0) -> float:
     """that waits for a random delay"""
     ran = random.uniform(0, max_delay)
     await asyncio.sleep(ran)
